feat(phase8): 代码审查完成 — SQL安全+错误分类+测试扩展+0 clippy警告
Phase 8.3 Code Review (CR1-CR6):
- SQL注入防护: ToSql单引号转义, QueryBuilder安全文档
- 错误处理: Error::unknown() → query_build/validation/relationship
- 边界测试: +8个测试(空model/错误分类/Display/列映射)
- 代码质量: #[must_use], Default for Schema, #[allow] lint修复
- clippy: nexus-data-orm + nexus-data-macros 0 warnings
- 测试: 41 passed; 0 failed

Phase 8.1-8.3 新增文件:
- 8.1 commons: method_name, prelude, tests
- 8.2 rdbc: extractor, ConnectionPool, MockDbClient, tests
- 8.3 orm: sea_orm bridge, connection, tests, model/query/relationships

Bilingual comments added to all existing modules.

+++ crates/nexus-data-rdbc/src/connection.rs
@@ -6,6 +6,8 @@
 //! This module provides database connection pooling and management.
 //! 本模块提供数据库连接池和管理。
 
+#![allow(dead_code)] // pub(crate) scaffolding items for future use; 内部脚手架，后续使用
+
 use crate::{DatabaseType, R2dbcError, R2dbcResult};
 use std::sync::Arc;
 use std::time::Duration;
@@ -319,8 +321,8 @@ impl ConnectionPool {
         let inner: Box<dyn PoolInner> = match database_type {
             DatabaseType::PostgreSQL => {
                 let sqlx_pool = sqlx::postgres::PgPoolOptions::new()
-                    .max_connections(config.max_size as u32)
-                    .min_connections(config.min_idle as u32)
+                    .max_connections(config.max_size)
+                    .min_connections(config.min_idle)
                     .acquire_timeout(config.connection_timeout)
                     .idle_timeout(config.idle_timeout)
                     .max_lifetime(config.max_lifetime)
